Consistent validation with pydantic & vocpub (disallow None)
None is not allowed according to the vocpub profile but the pydantic
model accepted it. The pydantic model is changed to only allow dates.

This also changes the default version from None to "automatic" which
conveys more clearly what is happening (version is automatically
derived and added in merge CI action of vocabulary).